requirements.txt
# Start of Selection
Flask
google-cloud-secret-manager
gunicorn
# End of Selection

